Remove unnecessary `\` escape of `@`
Note that some `@`s do need to be escaped.

By only escaping some `@`s, the patterns are more easily copied into
Visual Studio Code which objects to escaped `@`s.

The downside is that any future `@`s need to be carefully considered,
instead of a blanket rule of "always `\` escape `@`s".