Add census-5 dictionary
census-5 is based on a survey of repositories that were scanned with
check-spelling and collects tokens that were found in allow / expect for
each repository. If a token was present in at least five repositories,
then it's included in the dictionary.

See https://github.com/check-spelling-sandbox/census

There are additional dictionaries available:
- census:census-7.txt
- census:census-10.txt

They're strict subsets (only counting tokens that appear in at least
seven or ten repositories), but mostly downplay the value of the
census-5.txt dictionary.

As far as I can tell, the items in the census-5.txt dictionary are
reasonable even if they aren't technically "words". They're unlikely to
result in a large amount of false-negatives.